[strategies][policies] Move segment_intersection_info to intersection strategy.
In order to implement CS-specific segment_intersection_info picking
segments and assigning points differently for each CS.
