+++ include/boost/geometry/strategies/intersection_result.hpp
@@ -2,9 +2,8 @@
 
 // Copyright (c) 2007-2015 Barend Gehrels, Amsterdam, the Netherlands.
 
-// This file was modified by Oracle on 2015.
-// Modifications copyright (c) 2015 Oracle and/or its affiliates.
-
+// This file was modified by Oracle on 2015, 2016.
+// Modifications copyright (c) 2015-2016 Oracle and/or its affiliates.
 // Contributed and/or modified by Adam Wulkiewicz, on behalf of Oracle
 
 // Use, modification and distribution is subject to the Boost Software License,
@@ -72,19 +71,6 @@ struct segment_intersection_points
         : count(0)
     {}
 };
-
-// All assigned in cart_intersect, passed to intersection_points
-template <typename CoordinateType, typename PromotedType, typename SegmentRatio>
-struct segment_intersection_info
-{
-    typedef PromotedType promoted_type;
-
-    CoordinateType dx_a, dy_a;
-    CoordinateType dx_b, dy_b;
-    SegmentRatio robust_ra;
-    SegmentRatio robust_rb;
-};
-
 
 }} // namespace boost::geometry
